requirements.txt
requests==2.32.3
python-dotenv==1.0.1
apify-client>=1.9.0
apify-shared>=1.2.0
anthropic>=0.40.0
feedparser>=6.0.0
pillow>=10.0.0
